+++ napalm_ansible/modules/napalm_get_facts.py
@@ -3,8 +3,8 @@
 
 This file is part of Ansible
 
-Ansible is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
+Ansible is free software: you can redistribute it and/or modify it
+under the terms of the GNU General Public License as published by
 the Free Software Foundation, either version 3 of the License, or
 (at your option) any later version.
 
@@ -17,7 +17,16 @@ You should have received a copy of the GNU General Public License
 along with Ansible.  If not, see <http://www.gnu.org/licenses/>.
 """
 from __future__ import unicode_literals, print_function
-from ansible.module_utils.basic import AnsibleModule, return_values
+from ansible.module_utils.basic import AnsibleModule
+
+
+# FIX for Ansible 2.8 moving this function and making it private
+# greatly simplified for napalm-ansible's use
+def return_values(obj):
+    """ Return native stringified values from datastructures.
+
+    For use with removing sensitive values pre-jsonification."""
+    yield str(obj)
 
 
 DOCUMENTATION = '''
@@ -144,15 +153,6 @@ try:
     napalm_found = True
 except ImportError:
     pass
-
-# Legacy for pre-reunification napalm (remove in future)
-if not napalm_found:
-    try:
-        from napalm_base import get_network_driver     # noqa
-        from napalm_base import ModuleImportError      # noqa
-        napalm_found = True
-    except ImportError:
-        pass
 
 
 def main():