+++ napalm_ansible/modules/napalm_ping.py
@@ -14,7 +14,16 @@ You should have received a copy of the GNU General Public License
 along with Ansible.  If not, see <http://www.gnu.org/licenses/>.
 """
 from __future__ import unicode_literals, print_function
-from ansible.module_utils.basic import AnsibleModule, return_values
+from ansible.module_utils.basic import AnsibleModule
+
+
+# FIX for Ansible 2.8 moving this function and making it private
+# greatly simplified for napalm-ansible's use
+def return_values(obj):
+    """ Return native stringified values from datastructures.
+
+    For use with removing sensitive values pre-jsonification."""
+    yield str(obj)
 
 
 DOCUMENTATION = '''
@@ -136,15 +145,6 @@ try:
 except ImportError:
     pass
 
-# Legacy for pre-reunification napalm (remove in future)
-if not napalm_found:
-    try:
-        from napalm_base import get_network_driver   # noqa
-        from napalm_base import ModuleImportError    # noqa
-        napalm_found = True
-    except ImportError:
-        pass
-
 
 def main():
     module = AnsibleModule(
@@ -236,7 +236,7 @@ def main():
     except Exception as e:
         module.fail_json(msg="cannot close device connection: " + str(e))
 
-    module.exit_json(changed=False, results=ping_response)
+    module.exit_json(changed=False, ping_results=ping_response)
 
 
 if __name__ == '__main__':